requirements.txt
Django==2.0.7
djangorestframework==3.8.2
gunicorn==19.7.1
psycopg2==2.7.5
Pillow==5.2.0
django-cors-headers==2.1.0
